+++ js/main.js
@@ -1,4 +1,4 @@
-import {createPhotos} from './create-photos.js';
+import {createPhotos} from './data.js';
 
 const PHOTO_AMOUNT = 25;
